requirements.txt
google-cloud
pillow
pg8000